stores CongReg modifications in unsigned[]
this apparently makes life easier for clang, and if the compiler *can't*
figure out how to resolve a ConfReg commit, we're doing something wrong
anyway.

+++ include/hw/config_register.hpp
@@ -15,7 +15,7 @@ class ConfigurationRegister : public WritablePlatformRegister<RegisterName<true>
 
 	private:
 		RegisterName<false>* _target;
-		mutable uint8_t _status[ConfigurationRegister::words];
+		mutable unsigned _status[ConfigurationRegister::words];
 
 		enum {
 			WORD_LOADED = 1,
